requirements.txt
Pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
